Finished A couple of stuff
Finished a couple of methods such as the playCards method in the PLayer
Object, and also added bas variables in the Hero object. Created a new
class called CheckForAnything which will be used to put any effects in
effect, add any conditions to the currentTurn, and so on so forth.